+++ stroom-security-server/src/main/java/stroom/security/server/JWTService.java
@@ -161,7 +161,11 @@ public class JWTService implements HasHealthCheck {
             String usersEmail = authenticationServiceClients.newAuthenticationApi().verifyToken(token);
             return new AuthenticationToken(usersEmail, token);
         } catch (ApiException e) {
-            throw new RuntimeException("Unable to verify token remotely!", e);
+            String message = String.format(
+                    "Unable to verify token remotely! Message was: %s. HTTP response code was: %s. Response body was: %s",
+                    e.getMessage(), e.getCode(), e.getResponseBody());
+            LOGGER.debug(message);
+            throw new RuntimeException(message, e);
         }
     }